requirements.txt
subprocess.run
ipfshttpclient
google-cloud-storage
argparse
pymongo
tinydb